+++ test/src/Shovel-test.js
@@ -1,4 +1,5 @@
 import { Shovel } from '../../src/Shovel.js';
+import { Broker } from '../../src/Broker.js';
 
 describe('Shovel', () => {
   describe('ctor', () => {
@@ -8,6 +9,20 @@ describe('Shovel', () => {
 
     it('throws if name is not a string', () => {
       expect(() => new Shovel({})).to.throw(TypeError, /name/);
+    });
+  });
+
+  describe('.name', () => {
+    it('name cannot be changed', () => {
+      const source = new Broker();
+      source.assertExchange('events', 'topic');
+
+      const dest = new Broker();
+      dest.assertExchange('events', 'topic');
+
+      const shovel = new Shovel('spade', { broker: source, exchange: 'events' }, { broker: dest, exchange: 'events' });
+
+      expect(() => (shovel.name = 'my-name')).to.throw(TypeError);
     });
   });
 });